+++ test/controllers/reports_controller_test.rb
@@ -407,21 +407,4 @@ class ReportsControllerTest < ActionDispatch::IntegrationTest
 
     assert_select "button[disabled][aria-label=?]", I18n.t("reports.index.next_period")
   end
-
-  private
-    def capture_sql_queries
-      queries = []
-      callback = lambda do |_name, _started, _finished, _unique_id, payload|
-        next if payload[:cached]
-        next if %w[SCHEMA TRANSACTION].include?(payload[:name])
-
-        queries << payload[:sql].squish
-      end
-
-      ActiveSupport::Notifications.subscribed(callback, "sql.active_record") do
-        yield
-      end
-
-      queries
-    end
 end
