+++ test/support/sql_query_capture.rb
@@ -1,0 +1,17 @@
+module SqlQueryCapture
+  def capture_sql_queries
+    queries = []
+    callback = lambda do |_name, _started, _finished, _unique_id, payload|
+      next if payload[:cached]
+      next if %w[SCHEMA TRANSACTION].include?(payload[:name])
+
+      queries << payload[:sql].squish
+    end
+
+    ActiveSupport::Notifications.subscribed(callback, "sql.active_record") do
+      yield
+    end
+
+    queries
+  end
+end
